Update website's aesthetic and content
This commit is centered around refreshing the aesthetic of the website through the index.css file and improving the content in the index.html file. Changes to the CSS include altering the layout of some HTML elements to enhance the user experience and modifying some color schemes for a better overall look and feel. For the HTML, some sections have been rewritten to provide a more precise and informative description of the product. Further, the floating links class name was modified to "floating__links" in order to follow BEM naming conventions. This combats potential CSS specificity issues in the future.