Fix incorrect path in OG and Twitter image meta tags
Changed the file path for the images in the OG and Twitter meta tags within the index.html file. The original relative file path was "img/reproduction.png" and this was updated to the correct path "/img/reproduction.png". This ensures that the images load correctly when the site is shared on these platforms.