Update CSS for special class styling
Updated the CSS and SCSS files to adjust the styling of the `special` class. This change removes the box-shadow from the `special.mini` class and modifies the pseudo-elements used in the `special` class. This was done to enhance the UI appearance and to fix a styling issue where the box-shadow was not intended for the `special.mini` class. The pseudo-elements are now only applied to the `special` class items that are not of `mini` type.