Remove redundant JavaScript and CSS files
Deleted js/vendor/modernizr-3.11.2.min.js, js/plugins.js, mbt+.html, and css/index.css.map. The modernizr and plugins JavaScripts were removed as they were unused and were causing unnecessary bloat in the codebase. Similarly, the CSS map file was not needed for the current state of the application. Furthermore, the HTML file mbt+.html was also removed as it was not connected with the rest of the application.